user_service/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
flask-restx==1.3.0
flask-swagger-ui==4.11.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
Werkzeug==3.1.3
